refactor: split converter/metrics.go into records.go and session.go
- metrics.go keeps MetricsConverter, NewMetricsConverter, ParseRecords
- records.go: per-record filling functions (interpolateValue,
  fillPositionFromGPS, fillCadenceFromSteps, fillDistance,
  fillElevation, fillSpeedFromDistance, and their helpers)
- session.go: ParseSession plus computeCadenceInSpm/computeMaxSpeed
- Pure file reorganization, no signature or behavior changes
- metrics.go: 574 lines to 164
